% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{fft_magnitude_sliding}
\alias{fft_magnitude_sliding}
\title{Compute FFT magnitudes of a real-valued input slice.
Returns a Vec<f64> of magnitudes (length = input.len()).
#@export}
\usage{
fft_magnitude_sliding(input)
}
\description{
Compute FFT magnitudes of a real-valued input slice.
Returns a Vec<f64> of magnitudes (length = input.len()).
#@export
}
