% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{fft_parallel}
\alias{fft_parallel}
\title{Compute FFT magnitudes of a real-valued input slice.
Returns a Vec<f64> of magnitudes (length = input.len()).
# @export}
\usage{
fft_parallel(input)
}
\description{
Compute FFT magnitudes of a real-valued input slice.
Returns a Vec<f64> of magnitudes (length = input.len()).
# @export
}
