% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{fft_sliding_parallel}
\alias{fft_sliding_parallel}
\title{Parallel FFT over all windows}
\usage{
fft_sliding_parallel(signal, window_size, hop_size, threads)
}
\description{
Parallel FFT over all windows
}
