% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{generate_sine_mixture}
\alias{generate_sine_mixture}
\title{Parallel FFT over all windows}
\usage{
generate_sine_mixture(duration, sampling_rate, freqs, amps, phases, noise_sd)
}
\description{
Parallel FFT over all windows
}
