% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{r_generate_sine_mixture}
\alias{r_generate_sine_mixture}
\title{Generate a mixture of three sine waves}
\usage{
r_generate_sine_mixture(t, freqs, amps = c(1, 1, 1), phases = c(0, 0, 0))
}
\arguments{
\item{t}{Vector of time points}

\item{freqs}{A numeric vector of 3 frequencies (Hz)}

\item{amps}{A numeric vector of 3 amplitudes (default = c(1, 1, 1))}

\item{phases}{A numeric vector of 3 phases (radians, default = c(0, 0, 0))}
}
\value{
Numeric vector of the sine wave mixture
}
\description{
Generate a mixture of three sine waves
}
